netsecai-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.12
python-dotenv==1.0.1

# Packet analysis
scapy==2.6.1

# Async support
anyio==4.6.2

# AI
google-genai>=0.2.0
